// src/generated/graphql.ts
export type Maybe<T> = T | null;
export type InputMaybe<T> = Maybe<T>;
export type Exact<T extends { [key: string]: unknown }> = {
  [K in keyof T]: T[K];
};
// ...
export type Scalars = {
  ID: { input: string; output: string };
  String: { input: string; output: string };
  Boolean: { input: boolean; output: boolean };
  Int: { input: number; output: number };
  Float: { input: number; output: number };
  CountryCode: { input: any; output: any };
  FuzzyDateInt: { input: any; output: any };
  Json: { input: any; output: any };
};
// ...
export enum MediaSeason {
  /** Predominantly started airing between October and November */
  Fall = "FALL",
  /** Predominantly started airing between April and June */
  Spring = "SPRING",
  /** Predominantly started airing between July and September */
  Summer = "SUMMER",
  /** Predominantly started airing between January and March */
  Winter = "WINTER",
}
// ...
export enum MediaSource {
  /** Version 2+ only. Japanese Anime */
  Anime = "ANIME",
  /** Version 3 only. Comics excluding manga */
  Comic = "COMIC",
  /** Version 2+ only. Self-published works */
  Doujinshi = "DOUJINSHI",
  /** Version 3 only. Games excluding video games */
  Game = "GAME",
  /** Written work published in volumes */
  LightNovel = "LIGHT_NOVEL",
  /** Version 3 only. Live action media such as movies or TV show */
  LiveAction = "LIVE_ACTION",
  /** Asian comic book */
  Manga = "MANGA",
  /** Version 3 only. Multimedia project */
  MultimediaProject = "MULTIMEDIA_PROJECT",
  /** Version 2+ only. Written works not published in volumes */
  Novel = "NOVEL",
  /** An original production not based of another work */
  Original = "ORIGINAL",
  /** Other */
  Other = "OTHER",
  /** Version 3 only. Picture book */
  PictureBook = "PICTURE_BOOK",
  /** Video game */
  VideoGame = "VIDEO_GAME",
  /** Video game driven primary by text and narrative */
  VisualNovel = "VISUAL_NOVEL",
  /** Version 3 only. Written works published online */
  WebNovel = "WEB_NOVEL",
}

/** The current releasing status of the media */
export enum MediaStatus {
  /** Ended before the work could be finished */
  Cancelled = "CANCELLED",
  /** Has completed and is no longer being released */
  Finished = "FINISHED",
  /** Version 2 only. Is currently paused from releasing and will resume at a later date */
  Hiatus = "HIATUS",
  /** To be released at a later date */
  NotYetReleased = "NOT_YET_RELEASED",
  /** Currently releasing */
  Releasing = "RELEASING",
}
// ...
export type GetSeasonalAnimeQueryVariables = Exact<{
  season?: InputMaybe<MediaSeason>;
  year?: InputMaybe<Scalars["Int"]["input"]>;
  page?: InputMaybe<Scalars["Int"]["input"]>;
}>;

export type GetSeasonalAnimeQuery = {
  __typename?: "Query";
  Page?: {
    __typename?: "Page";
    pageInfo?: {
      __typename?: "PageInfo";
      hasNextPage?: boolean | null;
      total?: number | null;
    } | null;
    media?: Array<({ __typename?: "Media" } & MediaFragment) | null> | null;
  } | null;
};

export type MediaFragment = {
  __typename?: "Media";
  id: number;
  status?: MediaStatus | null;
  genres?: Array<string | null> | null;
  popularity?: number | null;
  episodes?: number | null;
  source?: MediaSource | null;
  averageScore?: number | null;
  siteUrl?: string | null;
  description?: string | null;
  title?: { __typename?: "MediaTitle"; romaji?: string | null } | null;
  coverImage?: { __typename?: "MediaCoverImage"; large?: string | null } | null;
  studios?: {
    __typename?: "StudioConnection";
    nodes?: Array<{
      __typename?: "Studio";
      id: number;
      name: string;
      siteUrl?: string | null;
    } | null> | null;
  } | null;
  airingSchedule?: {
    __typename?: "AiringScheduleConnection";
    nodes?: Array<{
      __typename?: "AiringSchedule";
      episode: number;
      airingAt: number;
    } | null> | null;
  } | null;
  nextAiringEpisode?: {
    __typename?: "AiringSchedule";
    episode: number;
    airingAt: number;
  } | null;
};